+++ src/greatest_prime_factor_table.rs
@@ -1,10 +1,10 @@
 use crate::least_prime_factor_table::least_prime_factor_table;
 
-pub fn greatest_prime_factor_table(size: usize) -> Vec<Option<u64>> {
+pub fn greatest_prime_factor_table(size: usize) -> Vec<Option<u32>> {
     let lpf = least_prime_factor_table(size);
     let mut gpf = vec![None; size];
     for i in 2..size {
-        gpf[i] = if lpf[i] == Some(i as u64) {
+        gpf[i] = if lpf[i] == Some(i as u32) {
             lpf[i]
         } else {
             gpf[i / lpf[i].unwrap() as usize]
